app/accounts/views.py
from rest_framework import viewsets

from accounts.models import Author
from accounts.serializers import AuthorSerializer


class AuthorViewSet(viewsets.ModelViewSet):
    queryset = Author.objects.all()
    serializer_class = AuthorSerializer
